Add parking area model
This adds the parking area model, migrations and admin.

The model has been designed to primarily accomodate the WFS at
kartta.hel.fi with the possibility of using other sources. There is an
unsolved issue of versioning changes in the data, depending on if the
areas will be changing a lot (which seems unlikely but is still
possible).

The admin uses OSM because the default map is unusable if one wants to
see if the data matches the real world (as in you might see that the
areas are in the correct city but not if they are on the right street).