Add parking area importer and management command
Adds a parking area importer and management command that is used to
populate the database with parking areas based on data from
kartta.hel.fi WFS.

All areas are treated as multipolygons so that there isn't any need for
creating more code for special cases.

The importer tries its best not to overwrite existing data and inform
users about what is happening.