Parking: Fix converting non-digit to int on save
When a Parking with a terminal_number number is saved, the save function
tries find a terminal by the terminal number.  This lookup does an
implicit conversion to integer, because the Terminal.number field is an
integer field.  That caused couple exceptions.

Fix this by making sure that only an integer or None is passed to the
filter call.