Add normalized_reg_num to Parking
Add a new field "normalized_reg_num" to Parking object and fill it in a
data migration.

This field will be used for looking up parkings by registration number,
but ignoring the dashes or spaces in them.  This is needed, because the
in data coming from the operators the same registration number can be
formatted as "ABC-123" or "ABC123".