Travis: Run tests with DEBUG=0
Tests should be run for production code, not development code, so set
DEBUG=0 when running tests.

There might be many differences between production and development code.
At least the following is known: In the development environment
"django_extensions" is added to INSTALLED_APPS, but it's not used in
production environment.