Add Region model with migrations
The dashboard will need also some bigger areas than the current "parking
areas".  Let's call these regions.

Add a new model Region with name, geometry and capacity_estimate fields.

Add a region field to Parking model so that it's fast to find parkings
of a region.

The capacity estimates of the regions will be calculated based on the
capacity estimates of the parking areas.  Add some methods to
ParkingAreaQuerySet to help calculating these estimates.