Add command for filling regions of parkings
New parkings with a location will get the region field filled on their
`save` method, but the existing parkings will have to be processed
separately.  Add a management command for that task.

The command divides the parkings to time intervals so that each time
interval has about 20000 parkings.  Then each time interval is processed
separately in a transaction.  This should (a) limit the amount of
resources the command consumes and (b) make it easy to abort the
command, if it takes too much resources, and then continue later where
it left off.