Update Prequ to 1.4.4 and Pip to 19.0.3
Update Prequ to 1.4.4 in the requirements check configured for Tox and
in the post-update hook script, which is ran by the deployment script.
In addition update Pip to the latest version, 19.0.3, in these same
files.

Remove "pip install pip==18.0" before_script command from the Travis
configuration, since it does not have any effect, because Tox will
create another virtualenv with a different Pip version for the
requirements check anyway.