Permits: Implement Permit model and APIs
Add models for Permit and PermitSeries.

Each permit belogns to a permit series.  Permit has an external_id which
can be used as a identifier in addition to the regular id.  The
external_id must be unique within a series.

Permit series have active flag.  The idea is that only one of the series
is active at any moment.

Add API endpoints to enforcement API for following actions:

 * creating and reading PermitSeries
 * activating a PermitSeries, which also deactivates the old one
 * creating, reading, updating and deleting Permits
 * creating and updating Permits in bulk
 * updating and deleting by external_id in the currently active series

Permits are using integer id rather than UUID because that is much
faster in the bulk creation