Update djangorestframework and django-filter
Update djangorestframework, djangorestframework-gis and django-filter to
newer versions.

Remove strictness specification from ValidParkingFilters in enforcement
and monitoring APIs, since those are no longer supported.

Remove the "name" parameter from django-filter Filter fields, because it
has been renamed, but filling the name is actually optional, since it
will be filled in by the meta class of the FilterSet.

Amend API timestamp parsing test, because the formatting of the error
has changed in DRF.