Permits: Normalize registration number in PermitCacheItem
Permit might be stored with a different format for the registration
number, e.g. with a dash or without, or even lower case letters.
Therefore the PermitCacheItems, which are used for querying the permits,
should be stored with the normalized registration number to make it
possible to find a permit even if the registration number is in
different format.