valid_parking: Fix old parking filtering
The valid_parking endpoint will return either valid parkings at the
specified time or if there is none, then it should return the latest
invalid parking as long as it's no older than grace period (15 mins).

This was implemented incorrectly, in the ValidParkingFilter.filter_time:
The filter should not use the queryset as an input parameter for the
filtering, as it was doing by testing if the filtering returned any
parkings (in "if valid_parkings" line), since it is not guarenteed in
which order the filters are applied and therefore we can't know if the
registaration number filter is already applied or not.  That would
affect the result of the "if valid_parkings" test inside the time
filter, because if the registration number filtering is not applied yet,
then there most likely always is some valid parking for any specified
time (as long as it is not in the future).

There was also another issue that the filter_time method did slicing and
if that filtering was applied before the registration number filter,
then the registration number filtering fails with

    AssertionError: Cannot filter a query once a slice has been taken.

This started to occur after updating the django-filter package so maybe
the ordering of the filter applying was more consistent with the old
version.  Relaying on it was still incorrect in the first place.  See
e.g. https://travis-ci.org/City-of-Helsinki/parkkihubi/jobs/538168733
for full back trace of the error.

Fix these issues by moving this "if no valid parkings were found, then
search within the grace duration" logic to the view set rather than
doing it in the filter set.