requirements: Use non-binary psycopg2
Build psycopg2 from sources, because there has been some issues with the
binary version and the compiled version is also more secure because it
links to system SSL libraries rather then bundling its own version of
them.  The problem with bundled version of libraries is to keep them
updated (especially for security fixes).