Improve validation and normalization of Permit JSON fields
Make sure that the data in the JSON fields is in correct format and
convert the timestamps to a normalized format before saving them.

Especially the timestamp comparisons needs the start and end times to be
in format that can be compared correctly.  ISO-8601 formatted string
which is always in same timezone works just fine for this.

Also update the factories, because the timestamps have to be generated
in the normalized format so that some of the tests won't fail because
they assume that the format of the stored timestamp is the same as the
timestamp format used in the request.