Improve model admins
Make admin module of PermitCacheItem read-only so that they cannot be
edited and the "permit" FK field won't be created as a select field
which would be populated with all the Permit objects, since that takes
very long time.

Use OSMGeoAdmin as the base class for the area model admins.