fix_geometries: Amend multipolygon fixing for Django 2
It seems that in Django 2.2 the geometry collections stored in to the
database get loaded as multipolygons, but still having a geometry
collection in their internal data.  Those will need to fixed in the same
way as GeometryCollection instances even though the outside class is a
MultiPolygon.

This makes the fix_geometries unit tests pass on Django 2.2.