Permits: Interpret timestamps without timezone as UTC
Interpret the non-timezoned timestamps entered into start_time and
end_time fields of the Permit objects as UTC, since we know that
currently the only client of this API is still pushing non-timezoned
timestamps.  This commit can be reverted when the client side code has
been updated to push timezoned timestamps.