check_parking: Handle GK25FIN projection failures
The WGS-84 location provided as the endpoint argument might be outside
of the GK25FIN projection plane and then an exception will be raised
when trying to transform the point to GK25FIN.  Handle that exception
gracefully and then just map the location to unknown payment zone and
permit area.

Add test cases which check that some extreme locations, which would
otherwise raise this error, are handled gracefully and even stored to
the database.