Travis: Fix codecov to work again
Currently codecov is not working, because there is an issue that it
doesn't work with the coverage 5.0 Python package.  Fix this issue by
making sure that coverage 4.x is used instead.

See the coverage error in the Travis build logs for latest master
builds (e.g. [1]) and the codecov issue [2]

[1] https://travis-ci.org/City-of-Helsinki/parkkihubi/jobs/639843555#L1003

[2] https://github.com/codecov/codecov-python/issues/224