Permitted user for PermitArea
Add permitted_user a FK to User table in order to identify the Permit
area the user has access to read.

Update 3 migrations to reflect the change:
1. 0029_enforcement_domain add foreign key to User table from
PermitArea with default null

2. 0030_domain_data update PermitArea's permitted_user with PASI user
as default

3. 0031_default_domain update PermitArea for not to accept default
value as null

Update PermitArea and Geojson importer to pass permitted-user as an
argument.

Update test_geojson_importing_management_commands and
test_importing_management_commands to pass PASI user to be used as
permitted_user for PermitArea