Operator: Add endpoint to activate permitseries
Add endpoint to activate permitseries. The id of permit series to
activate should be in url param. In the post request body, either
deactivate_others(Boolean) or deactivate_series(list of permit
series id to deactivate) should be specified. The id specified
in the url will be activated. The following is case based on
request body:
* deactivate_others is true: All other permit series will be
deactivated.
* deactivate_series: The permit series specified in
deactivate_series will be deactivated.