Move PermitSeries pruning to a classmethod
Move the permit series pruning code from activate API endpoint
implementation to a classmethod PermitSeries.delete_prunable_series.

This seems to be nicer place for the code, since it might have to be
used outside of the activate endpoint and moving it here makes the
code in the activate endpoint cleaner.