Travis,requirements: Fix coverage reporting
Pin the versions of codecov and coverage packages used by the last phase
in Travis configuration for sending coverage reports to codecov.io.

Make sure that the used version of covarage lib is the same for running
the tests and collecting coverage reports by adding the constraint line
"-c requirements-cov.txt" to requirements-test.in.