test_permit_model: Fix timestamp in test case
The start_time of the first subject might not be in any valid range of
the permit, since valid ranges is calculated by combining ranges of
subjects and areas. Think about a case where the permit is generated so
that all its areas have start_time later than the start_time of the
first subject.

This has caused several random test failures in CI.