factories,tests: Replace permit factories with functions
It is easier to control what kind of data is created when using regular
functions rather than the factories, so replace these permit factories
with regular functions.  This is important, because these tests must
create the Permit and PermitSeries objects with correct owner and domain
data.

Also, while converting the usage of factories to function calls, make
sure that the created permit series have correct owner and the created
permits have correct domain.

This change also reveals a bug in the Permit creation endpoint in the
Enforcement API: It should use the domain of the enforcer executing the
request, but it uses the default domain instead.  A few tests start to
fail because of this; mark those tests with xfail.  The fix will be
implemented in a following commit.