APIs: Fix Permit default domain logic
Only Enforcement API should set the domain when creating Permits, and in
Operator API the "domain" field is required.

When setting the domain in the Enforcement API, it should be set to the
enforced domain of the executing enforcer rather than the default
domain as it used to be.  Fix this bug and remove the xfail marks from
the tests which were affected by the bug.