Refactor PermitArea restriction
Move PermitArea usage validation from OperatorPermitSerializer to its
base class PermitSerializer so that same checks are also done for the
Enforcement API.  This code validates that PermitAreas used in a Permit
creation or modification are permitted for the user executing the
request.  Also rewrite it a bit while moving.

Fix the Enforcement API active_permit_by_external_id tests, which broke
because they tried to use unallowed PermitAreas.