Document pruning of old PermitSeries
Add a description to Operator and Enforcement API documents mentioning
that inactive permits series are pruned automatically after a certain
time.

Currently, the pruning is linked to activating new series, but this
might be changed at some point and is not considered relevant, so no
need to mention it in these endpoint descriptions.