Change PaymentZone PK from UUID to integer
This enables us to change it to foreignkey for parkings model later.

This commit is cherry-picked from commit a1b7f41bb in the multicity
branch to master to reorder the migrations so that most of them can be
reverted easily, since this one is not reversible, but the other
multicity migrations can be reverted.