valid_parking: Change zone field definition
Define a new field type CharOrIntegerField for the zone field to make it
clearer that the returned value could actually be either string or
integer and to avoid loading the PaymentZone object from the database.

Also add a few asserts to the related test case to make sure that the
created two zones are actually in different domains and that the
returned zone code is correct.