Change PermitArea PK from UUID to integer
This enables us to have PermitArea foreign-keys in PermitLookupItem
model without affecting the lookup performance too much.

This change was already done in the multicity branch (commit 4dc3be92f),
but is now reintroduced here in master to reorder the migrations so that
most of them can be reverted easily, since this one is not reversible,
but the other multicity migrations can be reverted.