Make 0034 migration reversible
Make it possible to run the 0034 migration backwards so that it can be
reverted, if there is any need to do so.

The AlterField added to 0033 migration is a hack to make the
PermitLookupItem.area_identifier field temporarily nullable just before
it is removed in 0034 migration.  That makes Django restore it as
nullable when 0034 is ran backwards, which is needed because it won't
have any values at that point.  The values are restored just after the
field is restored.