api/enforcement/valid_parking: Fix zone handling
The valid_parking endpoint was failing for Parkings which don't have a
zone, because it tried to access "zone.code" and that causes:

    AttributeError: 'NoneType' object has no attribute 'code'

when zone is None.

Fix this by simply populating the zone field in the to_representation
method and only doing it when the instance.zone has a value.

Also fix the DiscParkingFactory so that it generates Parking object
without a zone, since disc parkings don't have a zone and amend some
code in the tests accordingly:

  * Handle None zone in check_parking_data_matches_parking_object.

  * The test_endpoint_can_return_zone_as_string_or_integer cannot be
    tested with a disc parking because it doesn't have a zone, so
    remove the "parking_type" parametrization from it.