GeoJSON importers: Improve properties passing
Pass the properties dict as is to the defaults argument in
get_or_create.  This makes the PaymentZone importer work also for zones
that have non-integer domains, since previously it was requred to put
the code to a field named "number", but when that was passed in the
"defaults" argument of get_or_create, it resulted to an error, because
non-integer was tried to be saved to an integer field.