valid_permit_item: Do active filtering in base queryset
Make the base queryset in the ValidPermitItemViewSet limit to the active
permit items rather than doing it in each filter condition separately in
the ValidPermitItemFilter.  This way the active filtering is always done
only once, and there's no risk that it would be skipped if the filter
would allow leaving out both reg_num and time parameters in the future.