valid_permit_item: Paginate with CursorPagination
The pagination by page numbers doesn't work correctly with the
valid_permit_item endpoint.  It seems that sometimes a few items
"between the pages" were not returned when iterating all the pages.

Switch to more reliable cursor pagination and order the results with the
id field to make the pagination more reliable.

This way the first page always contains the newest items and when going
forward the pages, there shouldn't be any new items added between the
pages, because newer items have larger id number in the database.

The result schema will change a bit though, since with CursorPagination
there is no longer "count" field for the total number of items.  That
should't be a problem for the API users.  Update the documentation
accordingly.

While adding a custom pagination class, make the page size also a bit
bigger (from 100 to 200) and make it possible to request bigger page
size, up to 1000 items per page, with a "page_size" parameter.