Add make_batches queryset helper function
Add a helper function named `make_batches` which can be used to split a
queryset into batches of given number of objects.  The batches are
formed with filtering conditions based on given order-by field rather
than using the LIMIT and OFFSET SQL features (= slices in Django ORM),
because if LIMIT/OFFSET is used, then it's not possible to apply certain
ORM operations for the queryset anymore (e.g. taking the last item).