models/parking: Tune ArchivedParking db_indexes
Remove most of the database indexes of ArchivedParking table and add
db_index for the arhived_at field.

Having fewer database indexes in the ArchivedParking table will make
archiving operations faster, since DBMS doesn't have to update so many
indexes when inserting into the table.

These indexes are not needed, because archive is not used for searching.

The time_end index is kept and archived_at index is added, because they
are being used in the archiving process.

Removing the indexes makes the archiving process about 125% faster (or
2.25 times as fast as with indexes), based on these measurements:

 * with indexes (before this commit):   2000 items/s
 * without indexes (after this commit): 4500 items/s