settings: Allow setting TEST_DATABASE_TEMPLATE
When running the tests, Django will create a test database
automatically.  Parkkihubi needs this test database to have the PostGIS
extensions enabled, but it's possible that the default
template (template1) on the PostgreSQL instance doesn't have it.
Therefore it's useful to be able to define another template database for
the test database creation.

Another solution is to create the PostGIS extension to the template1,
but that's not always wanted, since it would also affect all other
databases created after that.