test_archive_parkings: Fix dates in test case
This test tries to check that the number of months given to the archiver
works correctly, but the test data that it generated was bad:

It generated 10 parkings with 60 days old end times and 10 parkings with
90 days old end times.  The idea seemed to be that the first 10 are
older than 1 month, but *not over 2 months old* and second 10 are older
than 2 months; so that running the archiving with keep-months=1 would
archive them all and with keep-months=2 would archive only the second
set.  However, if the current date D is from 2023-03-01 to 2023-04-28,
then the D - 60days is from 2022-12-31 to 2023-02-27, but D - 2months is
from 2023-01-01 to 2023-02-28.  This means that those D - 60days
parkings are actually over 2 months old and are therefore archived (as
they should be) when running the archiver with keep-months=2.  (The same
problem exists with the same dates on every year, except leap years.)

Fix this by using different end times for the created parkings so that
there really always is 10 parkings in the 1-2 month age range and 10
parkings over 2 months old.