Tune ends_before and created_before filters
The naming of these filters, "ends_before" and "created_before",
suggests that they shouldn't include objects which are ending or created
at exactly the given timestamp.  Fix the filter to use less-than
operator rather than less-than-or-equal operator to make the operation
match the naming.

These filters are only used in anonymization and archiving and changing
the meaning won't break neither of those features.