Move anonymize to QuerySet level
It's much faster to anonymize in bulk when it's done with a single SQL
query rather than looping over each object and calling anonymize
function.  Change the anonymization implementation to do that for
Parking, ArchivedParking and ParkingCheck models.

For PermitLookupItem model, the implementation has to do something else,
since the anonymization should also be done for the Permit object
holding those items.  It's still better to have the same API for that
model, so also move the anonymize function from PermitLookupItem to
PermitLookupItemQuerySet with needed modifications so that it works for
the queryset.

Note: The previous and current implementation of PermitLookupItem
anonymization does too much anonymization, since when anonymizing a
single item of a Permit, all items of that Permit are anonymized.  With
the current data set this is not a big problem, since most of the
Permits have just single subject.  However, there is a small percentage
of Permits which are anonymized incorrectly because of this and
therefore this problem should be addressed in the future.