test_archive_parkings: Check that Parkings have enough data
The Parking objects used in these tests are generated with factories and
therefore it's not clear that they have data present in all fields.

Add some checking that ensures that all of the fields have some data.

There is still four fields which have been commented out, since the
generated Parking objects have no data in those fields.  These will be
added to the check on next commit.