Add PermitAreaItem and PermitSubjectItem
Add new supporting models for storing Permit data: PermitAreaItem and
PermitSubjectItem.  These store the data from the "areas" and "subjects"
JSON fields in a format that is easier to search.  Even though
PostgreSQL allows making special queries to JSON fields, it's still
impossible (or at least I can't find how) to search for Permits that
have a non-empty registration number in the "subjects" field, but with
these new models it's easy.

The data to PermitSubjectItem and PermitAreaItem database tables will be
created automatically when Permit is created or updated (bulk creation
is also handled) and initially for the existing Permits it's generated
with a data migration (0044_populate_permit_items_from_json).

Additionally the PermitLookupItem objects get FKs to the subject_item
and area_item that it's originating from.