Fix Permit anonymization
Fix an issue in the permit anonymization, which caused permits to be
anonymized too soon.

 - The issue was that "anonymize_permit_registration_numbers" generated
   a queryset of all PermitLookupItems that have ended, but then all
   Permits related to those lookup items were anonymized totally even if
   the Permit still had some other items that have not ended.

Also improve performance of the permit anonymization by moving most of
the work from Python side to the DBMS.

The new implementation only anonymizes permits which are totally ended,
i.e. have no items that are still valid.  Another possibility is to
anonymize only those parts of the Permit that are no longer valid, but
that would much harder to implement.