CI: Run tests on containers
Use the "container" option to make the GitHub CI jobs to run on
containers so that we can use containers that support GDAL 2, since the
oldest image provided by GitHub directly is ubuntu-20.04 and it's
already using GDAL 3; and Django 2.2 is not compatible with GDAL 3.

The image used is python:VER-slim-buster (for ver in 3.5, 3.6, 3.8).

Add problem_matcher=False to tox.ini, because tox fail otherwise. See
https://github.com/ymyzk/tox-gh-actions/blob/v2.12.0/README.md#disable-problem-matchers