Dockerfile: Use separate pycache directory
The pyc files in the app directory are visible to the container when the
Git checkout is mounted as /app by Docker and if those pyc files are for
the same Python version, but are not compatible with the Python in the
container, it causes problems.  Make sure that the container reads the
pyc files only from inside the container.

This can also speed up the application inside the container, since now
it's possible to also write pyc files and keep them up-to-date, so
possibly some py -> pyc compiling could be cached instead of repeated on
every run.