run-pytest: Wait for database and use TTY if available
Wait for the database to be up before starting pytest.

Also improve the TTY handling, because if the pytest is run with
"docker-compose exec -T" in a terminal (TTY) and Ctrl+C is pressed, then
the output of the pytest command is just detached from the terminal, but
the command still continues to run in the container.  This is not good
and therefore we should just pass "-T" when there is no TTY available
(e.g. in the CI).