Add more complete Parking factories
Add CompleteParkingFactory and CompleteHistoryParkingFactory for
generating Parkings that have values in all fields.

Use those factories in test_archive_parkings so that we can be certain
that all fields are preserved when archiving.

For the "terminal" field add a new factory to generate ParkingTerminal
objects.

Make sure that the area generated for the parking areas and regions is
so big that it covers all the generated locations.  This ensures that
the Parking objects don't lose their "parking_area" or "region" values
on the Parking.save method, since it recalculates those.

Fix capacity_estimates to be lazy functions.