test_active_permit_by_external_id: Fix one test case
The test_operator_can_post_active_permit_by_external_id test case used
to pass by accident.  It generated a PermitArea which could have been
created with a different permitted_user as was used by the API client.
However, stars aligned correctly and the create_permit call (which was
done in the begining of the test) already created a PermitArea with the
same identifier and therefore that PermitArea had the correct
permitted_user (aka owner).

The recent change in the factories changed the star alignment and caused
this incorrectly written test case to start failing, since it changed
how the random data is consumed by the various faker calls.

The reason why the test case deterministically passed in the past and
now started to fail is that we use a constant seed for the fake library.
See the fake.seed call in parkings/factories/faker.py.

Fix the test case by defining the permitted_user for the generated area.