MethodOverrideMiddleware
Implement the method override feature as was documented in the previous
commit:

Make it possible to override the HTTP method of the API calls via a
query parameter named "method".  This allows using the API with some
limited clients which don't support all HTTP methods.

To enable this feature the real HTTP method has to be POST and (for now)
only supported methods are DELETE, PATCH and PUT.