Add and populate PermitArea.allowed_users
Add a new many-2-many field named "allowed_users" to PermitArea and
populate it with the user object from the "permitted_user" field.

This allows multiple operators/enforcers to use the same PermitAreas for
their permits.

also make the permitted_user field nullable for now to allow migrations
to be reversed.

Refs HPH-36