Dockerization: Add platform and change user uid
Specify platform as "linux/amd64" for the app container in the
docker-compose.yml, because that's what it's supposed to be and it won't
even work on e.g. ARM based archs. (Some deps won't compile.)

Change the appuser uid to 1000, since that is very often the same uid as
the developer's uid on a Linux host, which could avoid some problems
when the source tree is mounted as a volume to the container and still
being modified from the host.